# Conflicts:
#   src/main/java/ru/sirius/natayarik/ft/data/AccountDTO.java
#   src/main/java/ru/sirius/natayarik/ft/data/UserDTO.java
#   src/main/java/ru/sirius/natayarik/ft/services/OperationService.java

+++ src/main/java/ru/sirius/natayarik/ft/data/AccountDTO.java
@@ -13,7 +13,7 @@ public class AccountDTO extends BaseDTO {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
